scroot/autonomy/requirements.txt
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.25.0
sentencepiece>=0.1.99
safetensors>=0.3.1
pyserial>=3.5
